[SemanticVerifierPass]: fix and new tests.
We were iterating through the original statements and through the
CircuitGraph of the transformed program. If we did some reordering, our
analysis would say that the programs were not equal.
The answer is to do the oposite: build the circuit graph for the
original program and iterate through the transformed program's
instructions.