[InlineAllPass]: caching.
Due to slow inline times from big circuits, we decided cache the
inlinings so that we wouldn't have to do it one by one. Now, we don't
have a "loop until we have nothing more to inline", but a "loop through
each operation not in basis, and inline it with the cached
instructions".